+++ doc/releaseNotes.md
@@ -12,3 +12,10 @@ This file describes the main feature changes for each readout.exe released versi
 
 ## v0.19.1 - 07/03/2019
 - fix for FairMQ build.
+
+## next version
+- consumer-fileRecorder: added parameter pagesMax.
+- equipment-dummy: added a new data page filling pattern.
+- Added consumer-processor to allow data pages processing with configurable user-provided dynamic library.
+- consumer-*: added parameter consumerOutput, to chain consumers (e.g. to record data output from a consumer-processor instance).
+- Added libProcessorZlibCompress as example consumer-processor library.